Improved load.sh script

+++ scripts/load.sh
@@ -14,14 +14,21 @@ echo ""
 echo "Copying model files in" $runid "with label" $label "to the model input files."
 echo ""
 
-
 # Load hdf5 files
+
 for model in $runid/*_$label.h5
 do
-    input=${model/$runid\//}
-    input=${input/$label/input}
-    cp -v $model $input
+    if [[ -s $model ]]
+    then
+	input=${model/$runid\//}
+	input=${input/$label/input}
+	cp -v $model $input
+    else
+	echo "** Load error! " $model " does not exist!! **"
+    fi
+	echo ""
 done
+
 
 # load fort.44
 cp -v $runid/fort.44 fort.44
